# openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NextSessionDecision": {
        "description": "Output of CurriculumPlanner, scored by NSDQ.",
        "properties": {
          "blocks": {
            "items": {
              "$ref": "#/components/schemas/SessionBlock"
            },
            "title": "Blocks",
            "type": "array"
          },
          "deadline_status": {
            "$ref": "#/components/schemas/DeadlineStatus"
          },
          "deferred_concepts": {
            "items": {
              "type": "string"
            },
            "title": "Deferred Concepts",
            "type": "array"
          },
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "rationale": {
            "title": "Rationale",
            "type": "string"
          },
          "reviews_included": {
            "items": {
              "type": "string"
            },
            "title": "Reviews Included",
            "type": "array"
          },
          "session_date": {
            "format": "date",
            "title": "Session Date",
            "type": "string"
          },
          "total_minutes": {
            "minimum": 0.0,
            "title": "Total Minutes",
            "type": "integer"
          }
        },
        "required": [
          "session_date",
          "blocks",
          "total_minutes",
          "rationale",
          "deadline_status"
        ],
        "title": "NextSessionDecision",
        "type": "object"
      },
...
      "_AcceptedResponse": {
        "properties": {
          "accepted": {
            "title": "Accepted",
            "type": "boolean"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "accepted",
          "message"
        ],
        "title": "_AcceptedResponse",
        "type": "object"
      },
...
      "_AnswersRequest": {
        "properties": {
          "answers": {
            "items": {
              "$ref": "#/components/schemas/_Answer"
            },
            "title": "Answers",
            "type": "array"
          }
        },
        "required": [
          "answers"
        ],
        "title": "_AnswersRequest",
        "type": "object"
      },
...
      "_DashboardResponse": {
        "properties": {
          "deadline_status": {
            "$ref": "#/components/schemas/DeadlineStatus"
          },
          "estimated_sessions": {
            "title": "Estimated Sessions",
            "type": "integer"
          },
          "next_review": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Review"
          },
          "on_track": {
            "title": "On Track",
            "type": "boolean"
          },
          "progress": {
            "title": "Progress",
            "type": "number"
          },
          "projected_completion": {
            "format": "date",
            "title": "Projected Completion",
            "type": "string"
          },
          "strong": {
            "items": {
              "type": "string"
            },
            "title": "Strong",
            "type": "array"
          },
          "weak": {
            "items": {
              "type": "string"
            },
            "title": "Weak",
            "type": "array"
          },
          "why": {
            "title": "Why",
            "type": "string"
          }
        },
        "required": [
          "progress",
          "on_track",
          "deadline_status",
          "strong",
          "weak",
          "next_review",
          "why",
          "estimated_sessions",
          "projected_completion"
        ],
        "title": "_DashboardResponse",
        "type": "object"
      },
      "_DiagnosticResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AssessmentItem"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "_DiagnosticResponse",
        "type": "object"
      },
...
      "_GoalDetailsResponse": {
        "properties": {
          "concepts": {
            "items": {
              "$ref": "#/components/schemas/Concept"
            },
            "title": "Concepts",
            "type": "array"
          },
          "goal": {
            "$ref": "#/components/schemas/LearningGoal"
          },
          "learner_model": {
            "$ref": "#/components/schemas/LearnerModel"
          }
        },
        "required": [
          "goal",
          "concepts",
          "learner_model"
        ],
        "title": "_GoalDetailsResponse",
        "type": "object"
      },
      "_GoalRequest": {
        "properties": {
          "daily_minutes": {
            "maximum": 240.0,
            "minimum": 5.0,
            "title": "Daily Minutes",
            "type": "integer"
          },
          "deadline": {
            "format": "date",
            "title": "Deadline",
            "type": "string"
          },
          "preferred_formats": {
            "items": {
              "type": "string"
            },
            "title": "Preferred Formats",
            "type": "array"
          },
          "purpose": {
            "title": "Purpose",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "purpose",
          "deadline",
          "daily_minutes"
        ],
        "title": "_GoalRequest",
        "type": "object"
      },
      "_GoalResponse": {
        "properties": {
          "concepts": {
            "items": {
              "$ref": "#/components/schemas/Concept"
            },
            "title": "Concepts",
            "type": "array"
          },
          "goal": {
            "$ref": "#/components/schemas/LearningGoal"
          }
        },
        "required": [
          "goal",
          "concepts"
        ],
        "title": "_GoalResponse",
        "type": "object"
      },
      "_GoalUpdateRequest": {
        "properties": {
          "daily_minutes": {
            "anyOf": [
              {
                "maximum": 240.0,
                "minimum": 5.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Daily Minutes"
          },
          "paused": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paused"
          }
        },
        "title": "_GoalUpdateRequest",
        "type": "object"
      },
      "_GoalUpdateResponse": {
        "properties": {
          "goal": {
            "$ref": "#/components/schemas/LearningGoal"
          },
          "paused": {
            "title": "Paused",
            "type": "boolean"
          }
        },
        "required": [
          "goal",
          "paused"
        ],
        "title": "_GoalUpdateResponse",
        "type": "object"
      },
      "_LearnerResponse": {
        "properties": {
          "learner_model": {
            "$ref": "#/components/schemas/LearnerModel"
          }
        },
        "required": [
          "learner_model"
        ],
        "title": "_LearnerResponse",
        "type": "object"
      },
      "_RecommendationRejectionRequest": {
        "properties": {
          "reason": {
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "_RecommendationRejectionRequest",
        "type": "object"
      },
      "_ResearchRequest": {
        "properties": {
          "topic": {
            "title": "Topic",
            "type": "string"
          }
        },
        "required": [
          "topic"
        ],
        "title": "_ResearchRequest",
        "type": "object"
      },
      "_ResearchResponse": {
        "properties": {
          "sources": {
            "items": {
              "$ref": "#/components/schemas/Source"
            },
            "title": "Sources",
            "type": "array"
          }
        },
        "required": [
          "sources"
        ],
        "title": "_ResearchResponse",
        "type": "object"
      },
...
      "_SimulateDayRequest": {
        "properties": {
          "days": {
            "maximum": 365.0,
            "minimum": 1.0,
            "title": "Days",
            "type": "integer"
          }
        },
        "required": [
          "days"
        ],
        "title": "_SimulateDayRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Learning Manager API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/goals": {
      "post": {
        "operationId": "create_goal_goals_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/_GoalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/_GoalResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Goal"
      }
    },
    "/goals/{goal_id}": {
      "get": {
        "operationId": "get_goal_goals__goal_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "goal_id",
            "required": true,
            "schema": {
              "title": "Goal Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/_GoalDetailsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Goal"
      },
      "patch": {
        "operationId": "update_goal_goals__goal_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "goal_id",
            "required": true,
            "schema": {
              "title": "Goal Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/_GoalUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/_GoalUpdateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Goal"
      }
    },
    "/goals/{goal_id}/dashboard": {
      "get": {
        "operationId": "dashboard_goals__goal_id__dashboard_get",
        "parameters": [
          {
            "in": "path",
            "name": "goal_id",
            "required": true,
            "schema": {
              "title": "Goal Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/_DashboardResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Dashboard"
      }
    },
    "/goals/{goal_id}/diagnostic": {
      "post": {
        "operationId": "diagnostic_goals__goal_id__diagnostic_post",
        "parameters": [
          {
            "in": "path",
            "name": "goal_id",
            "required": true,
            "schema": {
              "title": "Goal Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/_DiagnosticResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Diagnostic"
      }
    },
    "/goals/{goal_id}/diagnostic/answers": {
      "post": {
        "operationId": "diagnostic_answers_goals__goal_id__diagnostic_answers_post",
        "parameters": [
          {
            "in": "path",
            "name": "goal_id",
            "required": true,
            "schema": {
              "title": "Goal Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/_AnswersRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/_LearnerResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Diagnostic Answers"
      }
    },
    "/goals/{goal_id}/next-session": {
      "get": {
        "operationId": "next_session_goals__goal_id__next_session_get",
        "parameters": [
          {
            "in": "path",
            "name": "goal_id",
            "required": true,
            "schema": {
              "title": "Goal Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "today",
            "required": true,
            "schema": {
              "format": "date",
              "title": "Today",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NextSessionDecision"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Next Session"
      }
    },
    "/goals/{goal_id}/recommendations/reject": {
      "post": {
        "operationId": "reject_recommendation_goals__goal_id__recommendations_reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "goal_id",
            "required": true,
            "schema": {
              "title": "Goal Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/_RecommendationRejectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/_AcceptedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reject Recommendation"
      }
    },
    "/goals/{goal_id}/research": {
      "post": {
        "operationId": "research_goals__goal_id__research_post",
        "parameters": [
          {
            "in": "path",
            "name": "goal_id",
            "required": true,
            "schema": {
              "title": "Goal Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/_ResearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/_ResearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Research"
      }
    },
    "/goals/{goal_id}/simulate-day": {
      "post": {
        "operationId": "simulate_day_goals__goal_id__simulate_day_post",
        "parameters": [
          {
            "in": "path",
            "name": "goal_id",
            "required": true,
            "schema": {
              "title": "Goal Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/_SimulateDayRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/_SimulateDayResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Simulate Day"
      }
    },
    "/goals/{goal_id}/sources": {
      "get": {
        "operationId": "sources_goals__goal_id__sources_get",
        "parameters": [
          {
            "in": "path",
            "name": "goal_id",
            "required": true,
            "schema": {
              "title": "Goal Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/_ResearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sources"
      }
    },
    "/sessions/{session_id}/alternative-explanation": {
      "post": {
        "operationId": "alternative_explanation_sessions__session_id__alternative_explanation_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/_AlternativeExplanationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Alternative Explanation"
      }
    },
    "/sessions/{session_id}/assess": {
      "post": {
        "operationId": "assess_sessions__session_id__assess_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/_AnswersRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/_AssessmentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Assess"
      }
    },
    "/sessions/{session_id}/feedback": {
      "post": {
        "operationId": "session_feedback_sessions__session_id__feedback_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/_SessionFeedbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/_AcceptedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Session Feedback"
      }
    }
  }
}